% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abseil.R
\docType{package}
\name{abseil-package}
\alias{abseil}
\alias{abseil-package}
\title{abseil: 'C++' Header Files and Compiled Library from 'Abseil'}
\description{
Wraps the 'Abseil' 'C++' library for use by R packages. Original files are
  from <https://github.com/abseil/abseil-cpp>. Patches are located at
  <https://github.com/doccstat/abseil-r/tree/main/local/patches>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://abseil.xingchi.li}
  \item \url{https://github.com/doccstat/abseil-r}
  \item Report bugs at \url{https://github.com/doccstat/abseil-r/issues}
}

}
\author{
\strong{Maintainer}: Xingchi Li \email{anthony.li@stat.tamu.edu} (\href{https://orcid.org/0009-0006-2493-0853}{ORCID}) [contributor, copyright holder]

Authors:
\itemize{
  \item Abseil Team [copyright holder]
}

}
\keyword{internal}
